desktop/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mechanic-desktop"
version = "0.2.1"
description = "Local companion tool for WoW addon development built on AFD"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "GPL-3.0" }
authors = [
    { name = "Falkicon" }
]
keywords = ["wow", "warcraft", "addon", "development", "afd"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "click>=8.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "watchfiles>=0.20",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# AFD integration (install separately or use this extra)
afd = [
    "afd[server,testing]>=0.1.0",
]
# MCP server support (for mech mcp command)
mcp = [
    "mcp>=1.0.0",
]
# Research commands (Gemini API)
research = [
    "google-genai>=0.8.0",
]
# Asset management (PNG to TGA conversion)
assets = [
    "Pillow>=10.0.0",
]
# All extras
all = [
    "mcp>=1.0.0",
    "google-genai>=0.8.0",
    "Pillow>=10.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/Falkicon/Mechanic"
Repository = "https://github.com/Falkicon/Mechanic"
Documentation = "https://github.com/Falkicon/Mechanic#readme"

[project.scripts]
mechanic = "mechanic.cli:main"
mech = "mechanic.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["mechanic", "mechanic.commands"]

